+++ myems-api/core/equipment.py
@@ -276,57 +276,17 @@ class EquipmentItem:
                                    title='API.BAD_REQUEST',
                                    description='API.THERE_IS_RELATION_WITH_COMBINED_EQUIPMENTS')
 
-        # check relation with meter
-        cursor.execute(" SELECT meter_id "
-                       " FROM tbl_equipments_meters "
-                       " WHERE equipment_id = %s ",
-                       (id_,))
-        rows_meters = cursor.fetchall()
-        if rows_meters is not None and len(rows_meters) > 0:
-            cursor.close()
-            cnx.close()
-            raise falcon.HTTPError(status=falcon.HTTP_400,
-                                   title='API.BAD_REQUEST',
-                                   description='API.THERE_IS_RELATION_WITH_METER')
+        # delete relation with meter
+        cursor.execute(" DELETE FROM tbl_equipments_meters WHERE equipment_id = %s ", (id_,))
 
-        # check relation with offline meter
-        cursor.execute(" SELECT offline_meter_id "
-                       " FROM tbl_equipments_offline_meters "
-                       " WHERE equipment_id = %s ",
-                       (id_,))
-        rows_offline_meters = cursor.fetchall()
-        if rows_offline_meters is not None and len(rows_offline_meters) > 0:
-            cursor.close()
-            cnx.close()
-            raise falcon.HTTPError(status=falcon.HTTP_400,
-                                   title='API.BAD_REQUEST',
-                                   description='API.THERE_IS_RELATION_WITH_OFFLINE_METER')
+        # delete relation with offline meter
+        cursor.execute(" DELETE FROM tbl_equipments_offline_meters WHERE equipment_id = %s ", (id_,))
 
-        # check relation with virtual meter
-        cursor.execute(" SELECT virtual_meter_id "
-                       " FROM tbl_equipments_virtual_meters "
-                       " WHERE equipment_id = %s ",
-                       (id_,))
-        rows_virtual_meters = cursor.fetchall()
-        if rows_virtual_meters is not None and len(rows_virtual_meters) > 0:
-            cursor.close()
-            cnx.close()
-            raise falcon.HTTPError(status=falcon.HTTP_400,
-                                   title='API.BAD_REQUEST',
-                                   description='API.THERE_IS_RELATION_WITH_VIRTUAL_METER')
+        # delete relation with virtual meter
+        cursor.execute(" DELETE FROM tbl_equipments_virtual_meters WHERE equipment_id = %s ", (id_,))
 
-        # check relation with commands
-        cursor.execute(" SELECT command_id "
-                       " FROM tbl_equipments_commands "
-                       " WHERE equipment_id = %s ",
-                       (id_,))
-        rows_commands = cursor.fetchall()
-        if rows_commands is not None and len(rows_commands) > 0:
-            cursor.close()
-            cnx.close()
-            raise falcon.HTTPError(status=falcon.HTTP_400,
-                                   title='API.BAD_REQUEST',
-                                   description='API.THERE_IS_RELATION_WITH_COMMANDS')
+        # delete relation with commands
+        cursor.execute(" DELETE FROM tbl_equipments_commands WHERE equipment_id = %s ", (id_,))
 
         # delete all associated parameters
         cursor.execute(" DELETE FROM tbl_equipments_parameters WHERE equipment_id = %s ", (id_,))