+++ CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 -
 ### Fixed
--
+- fixed issues in on_delete actions of equipment, shopfloor, store and tenant in myems-api
+- 
 ### Removed
 -